pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-stalker"
version = "0.2.0"
description = "Telegram bot that filters job vacancies with a web UI"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "10sorry", email = "rip.pip.pip@yandex.ru" },
]

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pyrogram",
  "aiosqlite",
  "python-dotenv",
  "aiohttp",
  "rich",
  "psutil",
  "jinja2",
  "pydantic",
  "pypdf",
  "python-docx",
  "beautifulsoup4",
]

[project.optional-dependencies]
gpu = [
  "pynvml; platform_system == 'Linux' or platform_system == 'Windows'",
  "pyamdgpuinfo; platform_system == 'Linux'",
]

[project.scripts]
job-stalker = "job_stalker.run:main"

[project.urls]
Homepage = "https://github.com/10sorry/JobStalker"

[tool.setuptools]
package-dir = {"" = "job-stalker/src"}
packages = ["job_stalker", "job_stalker.vacancy_parser"]
include-package-data = true

[tool.setuptools.package-data]
job_stalker = ["static-files/**/*", "templates-html/**/*"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
exclude = [
    "venv",
    "build",
    "dist",
    "__pycache__",
]

[[tool.mypy.overrides]]
module = [
    "pyrogram.*",
    "pynvml.*",
    "pyamdgpuinfo.*",
    "aiosqlite.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]
exclude = ["venv", "build", "dist"]
